requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.11
et-xmlfile==1.1.0
greenlet==1.1.2
idna==3.3
numpy==1.22.2
openpyxl==3.0.9
PyMySQL==1.0.2
python-dotenv==0.19.2
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.31
timezonefinder==5.2.0
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8